+++ api-gateway/internal/domain/ports/clients.go
@@ -12,6 +12,7 @@ type AuthClient interface {
 	SendRegisterRequest(ctx context.Context, email string, password string, role models.UserRole) (*models.Tokens, error)
 	SendLoginRequest(ctx context.Context, email string, password string) (*models.Tokens, error)
 	SendChangePasswordRequest(ctx context.Context, access string, old string, new string) error
+	SendRefreshTokensRequest(ctx context.Context, refresh string) (*models.Tokens, error)
 
 	Close()
 }
